+++ packages/ui/src/index.ts
@@ -5,3 +5,9 @@ export {
   type ElButtonVariant,
 } from './lib/button/button';
 export { ElLabel, type ElLabelVariant } from './lib/label/label';
+export {
+  ElSegmentedButton,
+  ElSegmentedButtonItem,
+  type ElSegmentedButtonSize,
+  type ElSegmentedButtonVariant,
+} from './lib/segmented-button/segmented-button';